refactor(cli): check tauri version on migration
